// src/module/chess/types/Coordinates.ts
export interface Coordinates {
    char: string;
    num: number;
}

export enum CharIndex {
    a = 1,
    b = 2,
    c = 3,
    d = 4,
    e = 5,
    f = 6,
    g = 7,
    h = 8
}

export const KeyIndex: any = {
    a: CharIndex.a,
    b: CharIndex.b,
    c: CharIndex.c,
    d: CharIndex.d,
    e: CharIndex.e,
    f: CharIndex.f,
    g: CharIndex.g,
    h: CharIndex.h
};